Fixed wording on dev_setup script, fixed some odd choices in the GPS lib
The wording for setting the install path made it look like a yes/no
question. This has been fixed.

Removed the mReady member variable from the GlobalPositioningSystem
class as the conditional variable already does that job. Not sure what
was going through my head.

Changed some formatting in the code.