GPS api changes, initial SPI and ADC implementations
The GPS class is no more. I didn't like the idea of having another
thread sitting in the background queing results to be consumed by
another thread. It would make it hard to syncronize GPS samples to other
samples that we may want some sort of time base for (like ADC, pressure)

IGpsProtocol is now the only GPS interface and it provides the facility
for getting GPS results. getResult is intended to be a non-blocking
call, returning true when there is valid data, and false otherwise.

SkyTraqBinaryProtocol no longer takes a SkyTraqVenus, it purely deals
with decoding SkyTraq binary messages. SkyTraqVenus is still around, but
it is now expected that setting the GPS into the proper mode is done at
a higher level, see example_gps.

The first pass attempt at the SPI driver is done. The central
SerialPeripheralInterface object exists mostly to manage allocation of
SPI channels. SPI channels are unique_ptrs to restrict ownership of the
channel. This dissallows different parts of the code from accessing SPI
channels that they probably have no buisness accessing. This works
because there will only ever be a single device on an SPI channel
(chip select)

The first pass attempt at the ADC driver is done. This takes in an SPI
channel, and a bitset containing the desired ADC channels to enable.
The ADC object then can give out AnalogInputs. AnalogInputs provide a
read-only mechanism to read the latest aquired ADC value for a given
channel. These inputs are given out as references and can be distributed
anywhere neccessary.

CMake project name changed to pifly from libpyfly so that the library
doesn't come out as liblibpifly.

Changed bracked styling because it has grown on me.